+++ src/dwistring.c
@@ -1,18 +1,15 @@
 /**
  */
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <stdarg.h>
+
 #include "dwistring.h"
 
-char* concatAll(const int num, ...){
+char* concatAll(const uint32_t num, ...){
 	va_list argp;
 	char* result = NULL;
 	char *temp = NULL;
 	char *temp2 = NULL;
-	int i = 0;
+	unsigned int i = 0;
 	va_start(argp, num);
 	for(i=0;i<num;i++){
 		temp = va_arg(argp, char *);
@@ -37,6 +34,6 @@ char* concat(const char* first, const char* second){
 }
 
 
-int getStringLength(const char* string){
+size_t getStringLength(const char* string){
 	return string != NULL ? strlen(string) : 0;
 }
